Fix Python 2 division and move acm test
